08.12 - final refactorings (step 2 - refactor the auth feature) - create an auth package. move all the auth related code to the auth package
