+++ src/main/java/com/jparkkennaby/store/auth/Jwt.java
@@ -1,4 +1,4 @@
-package com.jparkkennaby.store.services;
+package com.jparkkennaby.store.auth;
 
 import java.util.Date;
