+++ src/main/java/com/jparkkennaby/store/auth/JwtAuthenticationFilter.java
@@ -1,4 +1,4 @@
-package com.jparkkennaby.store.filters;
+package com.jparkkennaby.store.auth;
 
 import java.io.IOException;
 import java.util.List;
@@ -9,8 +9,6 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
 import org.springframework.stereotype.Component;
 import org.springframework.web.filter.OncePerRequestFilter;
-
-import com.jparkkennaby.store.services.JwtService;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
