+++ src/main/java/com/jparkkennaby/store/auth/JwtConfig.java
@@ -1,4 +1,4 @@
-package com.jparkkennaby.store.config;
+package com.jparkkennaby.store.auth;
 
 import javax.crypto.SecretKey;
